Refactor Autocomplete field resolving #153
